requirements.txt
pyjwt
plac
requests
